+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Data Logger: A new plot type for displaying live data as a growing scatter plot (#3202, #3205)
 * Generate: Created a `Generate.RandomWalker` class for producing an infinite amount of streaming random data
 * Ticks: Improved support for multi-line tick labels on vertical axes (#3208) _Thanks @raburton_
+* Text: Exposed `FontName` and `LabelText` properties
 
 ## ScottPlot 5.0.15
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-01-14_